first minishell version
