--- main.c
@@ -1,11 +1,0 @@
-int main(int argc, char *argv[])
-{
-	// gestionar "ls |"
-
-
-
-	// LEXER
-	// PARSER (CMD TABLE)
-	// CLEAN LEXER
-
-}
